Continue value type alterations to correct invalid cast errors to the SetLevel effect to get it fully functioning.

+++ Modules/Effect/SetLevel/SetLevelData.cs
@@ -13,7 +13,7 @@ namespace VixenModules.Effect.SetLevel
 	class SetLevelData : ModuleDataModelBase
 	{
 		[DataMember]
-		public float level { get; set; }
+		public double level { get; set; }
 
 		[DataMember]
 		public RGB color { get; set; }
